Rename parsl.providers.error to errors, plural, for consistency
