+++ parsl/providers/ad_hoc/ad_hoc.py
@@ -5,7 +5,7 @@ import time
 from parsl.channels import LocalChannel
 from parsl.launchers import SimpleLauncher
 from parsl.providers.base import ExecutionProvider, JobStatus, JobState
-from parsl.providers.error import ScriptPathError
+from parsl.providers.errors import ScriptPathError
 from parsl.utils import RepresentationMixin
 
 logger = logging.getLogger(__name__)
